add checkpoint_save and checkpoint_restore! functions
